Refactor python tests to that they automatically find test dirs.

+++ test/clean.sh
@@ -1,13 +1,16 @@
 #!/bin/bash
 
-for dir in MovHandlerTests . BasicInstructionHandlerTests LeaHandlerTests ControlFlowHandlerTests
+for dir in */
 do
-  # create dirs
-  mkdir -p "$dir"/obj
-  mkdir -p "$dir"/logs
+  if [[ $dir =~ "Tests" ]]; then
+      # create dirs
+      mkdir -p "$dir"/obj
+      mkdir -p "$dir"/logs
 
-  # remove artifacts from last test run
-  rm "$dir"/obj/* > /dev/null 2>&1
-  rm "$dir"/logs/* > /dev/null 2>&1
-  rm -r "$dir"/peasoup_executable_directory* > /dev/null 2>&1
+      # remove artifacts from last test run
+      rm "$dir"/obj/* > /dev/null 2>&1
+      rm "$dir"/logs/* > /dev/null 2>&1
+      rm -r "$dir"/peasoup_executable_directory* > /dev/null 2>&1
+  fi
+
 done
